+++ RELEASE_NOTES.md
@@ -23,3 +23,4 @@
 - Improved scheduler performance
 - Added additional stats collection and publishing for the scheduler component
 - Data memory management update for better performance
+- Added functions `iot_data_list_head_push_list` and `iot_data_list_tail_push_list` to support list merging